FinFlow/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.5
django-allauth==0.52.0
django-bootstrap4==22.3
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
protobuf==3.20.1
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
whitenoise==6.3.0
django-cors-headers==4.3.1
